src/python/vizwiz_vqa_clip-GPT_V1_retrain/requirements.txt
# requirements.txt

git+https://github.com/openai/CLIP.git
matplotlib
numpy
pandas
torch
scikit-learn
transformers
ftfy
tqdm
Pillow